refactor extract class cliente into multiple classes
