+++ src/main/java/cliente/ClienteEscritor1.java
@@ -1,0 +1,9 @@
+package cliente;
+
+import java.io.IOException;
+
+public class ClienteEscritor1 {
+    public static void main(String[] args) throws IOException {
+        ClienteEscritorBase.main(new String[0]);
+    }
+}
