+++ src/main/java/cliente/ClienteReceptor1.java
@@ -1,0 +1,9 @@
+package cliente;
+
+import java.io.IOException;
+
+public class ClienteReceptor1 {
+    public static void main(String[] args) throws IOException {
+        ClienteReceptorBase.main(new String[0]);
+    }
+}
